+++ client/src/test/suite/cfnlint.test.ts
@@ -2,23 +2,16 @@
 import * as vscode from 'vscode';
 import * as assert from 'assert';
 import * as fs from 'fs';
-import { getDocUri, activate, activateAndPreview, getDocPath } from './helper';
+import { getDocUri, activate, activateAndPreview, getDocPath, sleep, testDiagnostics } from './helper';
 
-suite('A test to create some better spacing for activation', () => {
-	const docUri = getDocUri('good.yaml');
-
-	test('A more elaborate sleep', async () => {
-		await activate(docUri);
-
-		const actualDiagnostics = vscode.languages.getDiagnostics(docUri);
-		console.log(actualDiagnostics);
-	});
-});
 
 suite('Should have failures with a bad template', () => {
 	const docUri = getDocUri('bad.yaml');
 
 	test('Diagnose bad template', async () => {
+		await activate(docUri);
+        await sleep(2000); // Wait for the diagnostics to compute on this file
+
 		await testDiagnostics(docUri, [
 			{
 				severity: vscode.DiagnosticSeverity.Error,
@@ -43,6 +36,8 @@ suite('Should not have failures on a good template', () => {
 	const docUri = getDocUri('good.yaml');
 
 	test('Diagnose good template', async () => {
+		await activate(docUri);
+        await sleep(2000); // Wait for the diagnostics to compute on this file
 		await testDiagnostics(docUri, []);
 	});
 });
@@ -51,6 +46,8 @@ suite('Should not have failures a non CloudFormation Template', () => {
 	const docUri = getDocUri('not_template.yaml');
 
 	test('Diagnose good template', async () => {
+		await activate(docUri);
+        await sleep(2000); // Wait for the diagnostics to compute on this file
 		await testDiagnostics(docUri, []);
 	});
 });
@@ -59,6 +56,8 @@ suite('Should have failures even though AWSTemplateFormatVersion isn\'t in the f
 	const docUri = getDocUri('still_a_template.yaml');
 
 	test('Diagnoses a bad template without AWSTemplateFormatVersion', async () => {
+		await activate(docUri);
+        await sleep(2000); // Wait for the diagnostics to compute on this file
 		await testDiagnostics(docUri, [
 			{
 				severity: vscode.DiagnosticSeverity.Error,
@@ -73,6 +72,8 @@ suite('Should have failures even though AWSTemplateFormatVersion isn\'t in the f
 	const docUri = getDocUri('still_a_template_2.yaml');
 
 	test('Diagnoses a bad template without AWSTemplateFormatVersion', async () => {
+		await activate(docUri);
+        await sleep(2000); // Wait for the diagnostics to compute on this file
 		await testDiagnostics(docUri, [
 			{
 				severity: vscode.DiagnosticSeverity.Error,
@@ -87,6 +88,8 @@ suite('Should have failures even with a space in the filename', () => {
 	const docUri = getDocUri('a template.yaml');
 
 	test('Diagnoses a bad template with spaces in the name', async () => {
+		await activate(docUri);
+        await sleep(2000); // Wait for the diagnostics to compute on this file
 		await testDiagnostics(docUri, [
 			{
 				severity: vscode.DiagnosticSeverity.Error,
@@ -121,19 +124,5 @@ function toRange(sLine: number, sChar: number, eLine: number, eChar: number) {
 	const start = new vscode.Position(sLine, sChar);
 	const end = new vscode.Position(eLine, eChar);
 	return new vscode.Range(start, end);
-}
-
-async function testDiagnostics(docUri: vscode.Uri, expectedDiagnostics: vscode.Diagnostic[]) {
-	await activate(docUri);
-
-	const actualDiagnostics = vscode.languages.getDiagnostics(docUri);
-	assert.equal(actualDiagnostics.length, expectedDiagnostics.length);
-
-	expectedDiagnostics.forEach((expectedDiagnostic, i) => {
-		const actualDiagnostic = actualDiagnostics[i];
-		assert.equal(actualDiagnostic.message, expectedDiagnostic.message);
-		assert.deepEqual(actualDiagnostic.range, expectedDiagnostic.range);
-		assert.equal(actualDiagnostic.severity, expectedDiagnostic.severity);
-	});
 }
