+++ client/src/test/suite/index.ts
@@ -20,7 +20,7 @@ export function run(): Promise<void> {
 			files.forEach(f => mocha.addFile(path.resolve(testsRoot, f)));
 
 			try {
-				mocha.timeout(10 * 1000);
+				mocha.timeout(50000);
 				// Run the mocha test
 				mocha.run(failures => {
 					if (failures > 0) {
